py_requirement.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
nest_asyncio
fastapi_sessions
sqlalchemy
python-jose[cryptography]
passlib[argon2]